Added new package test_data, text_format file test_timetable_data.py moved to test_data file timetable_text.py moved to text_format
